initial cutover of core to celery
