+++ nautobot/core/celery.py
@@ -5,9 +5,9 @@ import nautobot
 from celery import Celery
 
 # Set the default Django settings module for the 'celery' program.
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'nautobot.core.settings')
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "nautobot.core.settings")
 
-app = Celery('nautobot')
+app = Celery("nautobot")
 
 # Using a string here means the worker doesn't have to serialize
 # the configuration object to child processes.
@@ -15,12 +15,7 @@ app = Celery('nautobot')
 #   should have a `CELERY_` prefix.
 # First, we need to call setup on the app to initialize settings
 nautobot.setup()
-app.config_from_object('django.conf:settings', namespace='CELERY')
+app.config_from_object("django.conf:settings", namespace="CELERY")
 
 # Load task modules from all registered Django apps.
 app.autodiscover_tasks()
-
-
-@app.task(bind=True)
-def debug_task(self):
-    print(f'Request: {self.request!r}')
