+++ nautobot/extras/datasources/git.py
@@ -13,6 +13,7 @@ from django.db import transaction
 from django_rq import job
 import yaml
 
+from nautobot.core.celery import app
 from nautobot.dcim.models import Device, DeviceRole, Platform, Region, Site
 from nautobot.extras.choices import LogLevelChoices, JobResultStatusChoices
 from nautobot.extras.models import (
@@ -49,7 +50,7 @@ def enqueue_pull_git_repository_and_refresh_data(repository, request):
     )
 
 
-@job("default")
+@app.task()
 def pull_git_repository_and_refresh_data(repository_pk, request, job_result):
     """
     Worker function to clone and/or pull a Git repository into Nautobot, then invoke refresh_datasource_content().
