+++ nautobot/extras/jobs.py
@@ -28,6 +28,7 @@ from .forms import JobForm
 from .models import GitRepository
 from .registry import registry
 
+from nautobot.core.celery import app
 from nautobot.ipam.formfields import IPAddressFormField, IPNetworkFormField
 from nautobot.ipam.validators import (
     MaxPrefixLengthValidator,
@@ -789,7 +790,7 @@ def get_job(class_path):
     return jobs.get(grouping_name, {}).get(module_name, {}).get("jobs", {}).get(class_name, None)
 
 
-@job("default")
+@app.task()
 def run_job(data, request, job_result, commit=True, *args, **kwargs):
     """
     Helper function to call the "run()", "test_*()", and "post_run" methods on a Job.
