+++ nautobot/extras/signals.py
@@ -106,11 +106,11 @@ def handle_cf_removed_obj_types(instance, action, pk_set, **kwargs):
     """
     if action == "post_remove":
         # Existing content types have been removed from the custom field, delete their data
-        delete_custom_field_data.delay(instance.name, pk_set)
+        delete_custom_field_data.delay(instance.name, list(pk_set))
 
     elif action == "post_add":
         # New content types have been added to the custom field, provision them
-        provision_field.delay(instance.pk, pk_set)
+        provision_field.delay(instance.pk, list(pk_set))
 
 
 m2m_changed.connect(handle_cf_removed_obj_types, sender=CustomField.content_types.through)
