+++ nautobot/extras/tasks.py
@@ -1,14 +1,20 @@
+import hashlib
+import hmac
 from logging import getLogger
 
+import requests
+from django.conf import settings
 from django.contrib.contenttypes.models import ContentType
-from django_rq import job
+from jinja2.exceptions import TemplateError
 
-from nautobot.extras.choices import CustomFieldTypeChoices
+from nautobot.core.celery import app
+from nautobot.extras.choices import CustomFieldTypeChoices, ObjectChangeActionChoices
+
 
 logger = getLogger("nautobot.extras.tasks")
 
 
-@job("custom_fields")
+@app.task()
 def update_custom_field_choice_data(field_id, old_value, new_value):
     """
     Update the values for a custom field choice used in objects' _custom_field_data for the given field.
@@ -49,7 +55,7 @@ def update_custom_field_choice_data(field_id, old_value, new_value):
         return False
 
 
-@job("custom_fields")
+@app.task()
 def delete_custom_field_data(field_name, content_type_pk_set):
     """
     Delete the values for a custom field
@@ -65,7 +71,7 @@ def delete_custom_field_data(field_name, content_type_pk_set):
             obj.save()
 
 
-@job("custom_fields")
+@app.task()
 def provision_field(field_id, content_type_pk_set):
     """
     Provision a new custom field on all relevant content type object instances.
@@ -87,3 +93,86 @@ def provision_field(field_id, content_type_pk_set):
         for obj in model.objects.all():
             obj._custom_field_data[field.name] = field.default
             obj.save()
+
+
+@app.task()
+def process_webhook(webhook_pk, data, model_name, event, timestamp, username, request_id):
+    """
+    Make a POST request to the defined Webhook
+    """
+
+    def generate_signature(request_body, secret):
+        """
+        Return a cryptographic signature that can be used to verify the authenticity of webhook data.
+        """
+        hmac_prep = hmac.new(key=secret.encode("utf8"), msg=request_body, digestmod=hashlib.sha512)
+        return hmac_prep.hexdigest()
+
+    from nautobot.extras.models import Webhook  # avoiding circular import
+
+    webhook = Webhook.objects.get(pk=webhook_pk)
+
+    context = {
+        "event": dict(ObjectChangeActionChoices)[event].lower(),
+        "timestamp": timestamp,
+        "model": model_name,
+        "username": username,
+        "request_id": request_id,
+        "data": data,
+    }
+
+    # Build the headers for the HTTP request
+    headers = {
+        "Content-Type": webhook.http_content_type,
+    }
+    try:
+        headers.update(webhook.render_headers(context))
+    except (TemplateError, ValueError) as e:
+        logger.error("Error parsing HTTP headers for webhook {}: {}".format(webhook, e))
+        raise e
+
+    # Render the request body
+    try:
+        body = webhook.render_body(context)
+    except TemplateError as e:
+        logger.error("Error rendering request body for webhook {}: {}".format(webhook, e))
+        raise e
+
+    # Prepare the HTTP request
+    params = {
+        "method": webhook.http_method,
+        "url": webhook.payload_url,
+        "headers": headers,
+        "data": body.encode("utf8"),
+    }
+    logger.info(
+        "Sending {} request to {} ({} {})".format(params["method"], params["url"], context["model"], context["event"])
+    )
+    logger.debug(params)
+    try:
+        prepared_request = requests.Request(**params).prepare()
+    except requests.exceptions.RequestException as e:
+        logger.error("Error forming HTTP request: {}".format(e))
+        raise e
+
+    # If a secret key is defined, sign the request with a hash of the key and its content
+    if webhook.secret != "":
+        prepared_request.headers["X-Hook-Signature"] = generate_signature(prepared_request.body, webhook.secret)
+
+    # Send the request
+    with requests.Session() as session:
+        session.verify = webhook.ssl_verification
+        if webhook.ca_file_path:
+            session.verify = webhook.ca_file_path
+        response = session.send(prepared_request, proxies=settings.HTTP_PROXIES)
+
+    if 200 <= response.status_code <= 299:
+        logger.info("Request succeeded; response status {}".format(response.status_code))
+        return "Status {} returned, webhook successfully processed.".format(response.status_code)
+    else:
+        logger.warning("Request failed; response status {}: {}".format(response.status_code, response.content))
+        raise requests.exceptions.RequestException(
+            "Status {} returned with content '{}', webhook FAILED to process.".format(
+                response.status_code, response.content
+            )
+        )
